requirements.txt
asgiref==3.6.0
certifi==2023.5.7
chardet==3.0.4
Click==7.0
Django==4.2.1
download-gitignore==2.0.1
idna==2.8
paho-mqtt==1.6.1
PyMySQL==1.0.3
requests==2.22.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.25.11
